requirements.txt
fastapi
pydantic
sqlalchemy
psycopg2
uvicorn
python-multipart
minio
pytest
requests
sqlalchemy-utils
bcrypt
passlib
pyjwt